Рефакторинг. Добавила интерфейс Phonebook, и заимплементила его в классах Phonebook1 и Phonebook2
