+++ lesson3/src/main/java/ru/geekbrains/alexkrasnova/javaleveltwo/lesson3/phonebook/impl/Phonebook2.java
@@ -1,21 +1,21 @@
-package ru.geekbrains.alexkrasnova.javaleveltwo.lesson3.phonebook;
+package ru.geekbrains.alexkrasnova.javaleveltwo.lesson3.phonebook.impl;
+
+import ru.geekbrains.alexkrasnova.javaleveltwo.lesson3.phonebook.Phonebook;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class Phonebook2 {
-    Map<String, List<String>> data;
+public class Phonebook2 implements Phonebook {
 
-/*    public Phonebook2(Map<String, List<String>> data) {
-        this.data = data;
-    }*/
+    Map<String, List<String>> data;
 
     public Phonebook2(){
         data = new HashMap<>();
     }
 
+    @Override
     public void add(String surname, String phoneNumber){
         List<String> phoneNumbers;
         if (data.containsKey(surname)){
@@ -27,10 +27,12 @@ public class Phonebook2 {
         data.put(surname, phoneNumbers);
     }
 
+    @Override
     public List<String> get(String surname){
         return data.get(surname);
     }
 
+    @Override
     public void print(){
         System.out.println(data);
     }
